Add proper first/last source code location fields/methods to tokens and nodes.

+++ include/common/sourcecodelocation.hpp
@@ -3,4 +3,4 @@
 #include <tuple>
 #include <cstddef>
 
-typedef std::tuple<size_t, size_t, size_t> SourceCodeLocation;
+typedef std::tuple<int, int, int> SourceCodeLocation;
